+++ integration/clients/Event.int.test.ts
@@ -1,5 +1,6 @@
 import 'whatwg-fetch'
 import { Event, EventKey, EventName, EventService, ObserveEvent, SystemEvent } from '../../src/clients/event'
+import { ObserveEventNames } from '../../src/clients/event/models/ObserveEventNames'
 import { setAppName } from '../../src/config'
 import { Done, Prefix, Subsystem } from '../../src/models'
 import { startServices, stopServices } from '../utils/backend'
@@ -7,16 +8,17 @@ import { startServices, stopServices } from '../utils/backend'
 jest.setTimeout(30000)
 
 let eventService: EventService
-
-beforeAll(async () => {
+beforeAll(() => {
   //todo: fix this console.error for jsdom errors
   console.error = jest.fn()
   setAppName('test-app')
+})
+beforeEach(async () => {
   await startServices(['Gateway'])
   eventService = await EventService()
 })
 
-afterAll(async () => {
+afterEach(async () => {
   await stopServices()
   jest.clearAllMocks()
 })
@@ -41,17 +43,17 @@ describe('Event Client', () => {
       const prefix = new Prefix('ESW', 'ncc.trombone2')
       const eventName = new EventName('offline')
       const eventKeys = new Set<EventKey>([new EventKey(prefix, eventName)])
-      const observeEvent = ObserveEvent.make(prefix, eventName, [])
+      const systemEvent = SystemEvent.make(prefix, eventName, [])
       expect.assertions(1)
 
       const callback = (event: Event) => {
-        expect(event).toEqual(observeEvent)
+        expect(event).toEqual(systemEvent)
         subscription.cancel()
         jestDoneCallback()
       }
 
-      const subscription = await eventService.subscribe(eventKeys, 1)(callback)
-      await eventService.publish(observeEvent)
+      const subscription = eventService.subscribe(eventKeys, 1)(callback)
+      await eventService.publish(systemEvent)
     })
   })
 
@@ -60,6 +62,24 @@ describe('Event Client', () => {
       const prefix = new Prefix('CSW', 'ncc.trombone')
       const eventName = new EventName('offline')
       const subsystem: Subsystem = 'CSW'
+      const systemEvent = SystemEvent.make(prefix, eventName, [])
+      expect.assertions(1)
+
+      const callback = (event: Event) => {
+        expect(event).toEqual(systemEvent)
+        subscription.cancel()
+        jestDoneCallback()
+      }
+
+      const subscription = eventService.pSubscribe(subsystem, 1, '.*')(callback)
+      await eventService.publish(systemEvent)
+    })
+  })
+
+  test('should subscribe to publish observe events | ESW-582', () => {
+    return new Promise<void>(async (jestDoneCallback) => {
+      const prefix = new Prefix('CSW', 'sequencer')
+      const eventName = ObserveEventNames.ObservationStart
       const observeEvent = ObserveEvent.make(prefix, eventName, [])
       expect.assertions(1)
 
@@ -69,7 +89,7 @@ describe('Event Client', () => {
         jestDoneCallback()
       }
 
-      const subscription = await eventService.pSubscribe(subsystem, 1, '.*')(callback)
+      const subscription = eventService.subscribeObserveEvents(1)(callback)
       await eventService.publish(observeEvent)
     })
   })