+++ integration/utils/shell.ts
@@ -36,8 +36,8 @@ const appLauncherSync = (name: string, version = 'master-SNAPSHOT') =>
 // Put SHA of 10 characters to make it consistent with Git api for latest SHA
 // This consistency avoids multiple jitpack triggers for same commit
 
-const csw_sha = 'de2df21'
-const esw_sha = '53a0071'
+const csw_sha = '5088284'
+const esw_sha = '06a6bf1'
 
 export const executeServicesScript = appLauncher('backend-testkit-services', esw_sha)
 export const executeComponentScript = appLauncher('backend-testkit-component', esw_sha)